requirements.txt
python-dotenv
openai
twilio
Flask